requirements.txt
aniso8601==3.0.2
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.13.1
flask-marshmallow==0.9.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
marshmallow==3.0.0b12
marshmallow-sqlalchemy==0.14.1
PyJWT==1.6.4
python-dotenv==0.9.1
pytz==2018.5
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24
Werkzeug==0.14.1
